Refactoring and changed index name
