+++ serverless/dynamoplus/repository/repositories.py
@@ -5,7 +5,7 @@ import logging
 from pynamodb.exceptions import DoesNotExist
 
 from dynamoplus.models.system.collection.collection import Collection
-from dynamoplus.models.indexes.indexes import Index, Query
+from dynamoplus.models.query.query import Index, Query
 from dynamoplus.repository.models import Model, IndexModel, QueryResult, IndexDataModel, DataModel, SystemDataModel
 from dynamoplus.utils.utils import convertToString, find_value, sanitize
