+++ serverless/test/dynamoplus/service/system/test_systemService.py
@@ -1,7 +1,7 @@
 import unittest
 import decimal
 
-from dynamoplus.models.indexes.indexes import Query
+from dynamoplus.models.query.query import Query
 from dynamoplus.repository.repositories import DynamoPlusRepository, IndexDynamoPlusRepository
 from dynamoplus.repository.models import Model, QueryResult
 from dynamoplus.service.system.system import SystemService
@@ -70,7 +70,7 @@ class TestSystemService(unittest.TestCase):
     @patch.object(DynamoPlusRepository, "create")
     @patch.object(DynamoPlusRepository, "__init__")
     def test_createIndexWithOrdering(self, mock_repository, mock_create):
-        expected_id = 'field1__field2.field21__ORDER_BY__field2.field21'
+        expected_id = 'example#field1__field2.field21__ORDER_BY__field2.field21'
         expected_conditions = ["field1", "field2.field21"]
         target_index = {"name": expected_id, "collection": {"name": "example"}, "conditions": expected_conditions,
                        "ordering_key": "field2.field21"}
@@ -87,7 +87,7 @@ class TestSystemService(unittest.TestCase):
     @patch.object(DynamoPlusRepository, "create")
     @patch.object(DynamoPlusRepository, "__init__")
     def test_createIndexWithNoOrdering(self, mock_repository, mock_create):
-        expected_id = 'field1__field2.field21'
+        expected_id = 'example#field1__field2.field21'
         expected_conditions = ["field1", "field2.field21"]
         target_index = {"name": expected_id, "collection": {"name": "example"}, "conditions": expected_conditions, "ordering_key": None}
         index_metadata = Collection("index", "name")
@@ -123,5 +123,5 @@ class TestSystemService(unittest.TestCase):
             [Model(Collection("index", "name"), {"name": "collection.name", "collection":{"name":"example"},"conditions":["collection.name"]}, False)])
         indexes = self.systemService.find_indexes_from_collection_name("example")
         self.assertTrue(len(indexes) == 1)
-        self.assertEqual(indexes[0].index_name, "collection.name")
+        self.assertEqual(indexes[0].index_name, "example#collection.name")
         self.assertEqual(call(expected_query), mock_find.call_args_list[0])
